+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file. This projec
 - New `forge make:block` command to easily scaffold a new block.
 - New `forge setup:block-build` command to easily scaffold the block build process.
 - Abstract `Job` class for creating long-running background tasks.
+- Abstract `Bulk_Action` class for quickly scaffolding new bulk actions.
 
 ### Fixed